split create explorer async and sync methods into different files
