+++ src/createExplorer.ts
@@ -1,80 +1,14 @@
 import path from 'path';
-import * as loaders from './loaders';
-import { readFile, readFileSync } from './readFile';
-import { cacheWrapper, cacheWrapperSync } from './cacheWrapper';
-import { getDirectory, getDirectorySync } from './getDirectory';
-import { getPropertyByPath } from './getPropertyByPath';
-import {
-  CosmiconfigResult,
-  ExplorerOptions,
-  LoaderEntry,
-  Config,
-  Cache,
-} from './types';
-import { LoaderAsync, LoaderSync } from './index';
+import { ExplorerBase, getExtensionDescription } from './createExplorerBase';
+import { readFile } from './readFile';
+import { cacheWrapper } from './cacheWrapper';
+import { getDirectory } from './getDirectory';
+import { CosmiconfigResult, ExplorerOptions, LoadedFileContent } from './types';
+import { LoaderAsync } from './index';
 
-// An object value represents a config object.
-// null represents that the loader did not find anything relevant.
-// undefined represents that the loader found something relevant
-// but it was empty.
-type LoadedFileContent = Config | null | undefined;
-
-class Explorer {
-  private readonly loadCache?: Cache;
-  private readonly loadSyncCache?: Cache;
-  private readonly searchCache?: Cache;
-  private readonly searchSyncCache?: Cache;
-  private readonly config: ExplorerOptions;
-
+class ExplorerAsync extends ExplorerBase<ExplorerOptions> {
   public constructor(options: ExplorerOptions) {
-    if (options.cache === true) {
-      this.loadCache = new Map();
-      this.loadSyncCache = new Map();
-      this.searchCache = new Map();
-      this.searchSyncCache = new Map();
-    }
-
-    this.config = options;
-    this.validateConfig();
-  }
-
-  public clearLoadCache(): void {
-    if (this.loadCache) {
-      this.loadCache.clear();
-    }
-    if (this.loadSyncCache) {
-      this.loadSyncCache.clear();
-    }
-  }
-
-  public clearSearchCache(): void {
-    if (this.searchCache) {
-      this.searchCache.clear();
-    }
-    if (this.searchSyncCache) {
-      this.searchSyncCache.clear();
-    }
-  }
-
-  public clearCaches(): void {
-    this.clearLoadCache();
-    this.clearSearchCache();
-  }
-
-  private validateConfig(): void {
-    const config = this.config;
-
-    config.searchPlaces.forEach((place): void => {
-      const loaderKey = path.extname(place) || 'noExt';
-      const loader = config.loaders[loaderKey];
-      if (!loader) {
-        throw new Error(
-          `No loader specified for ${getExtensionDescription(
-            place,
-          )}, so searchPlaces item "${place}" is invalid`,
-        );
-      }
-    });
+    super(options);
   }
 
   public async search(
@@ -82,13 +16,6 @@ class Explorer {
   ): Promise<CosmiconfigResult> {
     const startDirectory = await getDirectory(searchFrom);
     const result = await this.searchFromDirectory(startDirectory);
-
-    return result;
-  }
-
-  public searchSync(searchFrom: string = process.cwd()): CosmiconfigResult {
-    const startDirectory = getDirectorySync(searchFrom);
-    const result = this.searchFromDirectorySync(startDirectory);
 
     return result;
   }
@@ -116,30 +43,6 @@ class Explorer {
     return run();
   }
 
-  private searchFromDirectorySync(dir: string): CosmiconfigResult {
-    const absoluteDir = path.resolve(process.cwd(), dir);
-
-    const run = (): CosmiconfigResult => {
-      const result = this.searchDirectorySync(absoluteDir);
-      const nextDir = this.nextDirectoryToSearch(absoluteDir, result);
-
-      if (nextDir) {
-        return this.searchFromDirectorySync(nextDir);
-      }
-
-      const transformResult = this.config.transform(result);
-
-      // @ts-ignore
-      return transformResult;
-    };
-
-    if (this.searchSyncCache) {
-      return cacheWrapperSync(this.searchSyncCache, absoluteDir, run);
-    }
-
-    return run();
-  }
-
   private async searchDirectory(dir: string): Promise<CosmiconfigResult> {
     for await (const place of this.config.searchPlaces) {
       const placeResult = await this.loadSearchPlace(dir, place);
@@ -151,25 +54,6 @@ class Explorer {
 
     // config not found
     return null;
-  }
-
-  private searchDirectorySync(dir: string): CosmiconfigResult {
-    for (const place of this.config.searchPlaces) {
-      const placeResult = this.loadSearchPlaceSync(dir, place);
-
-      if (this.shouldSearchStopWithResult(placeResult) === true) {
-        return placeResult;
-      }
-    }
-
-    // config not found
-    return null;
-  }
-
-  private shouldSearchStopWithResult(result: CosmiconfigResult): boolean {
-    if (result === null) return false;
-    if (result.isEmpty && this.config.ignoreEmptySearchPlaces) return false;
-    return true;
   }
 
   private async loadSearchPlace(
@@ -184,48 +68,6 @@ class Explorer {
     return result;
   }
 
-  private loadSearchPlaceSync(dir: string, place: string): CosmiconfigResult {
-    const filepath = path.join(dir, place);
-    const content = readFileSync(filepath);
-
-    const result = this.createCosmiconfigResultSync(filepath, content);
-
-    return result;
-  }
-
-  private nextDirectoryToSearch(
-    currentDir: string,
-    currentResult: CosmiconfigResult,
-  ): string | null {
-    if (this.shouldSearchStopWithResult(currentResult)) {
-      return null;
-    }
-    const nextDir = nextDirUp(currentDir);
-    if (nextDir === currentDir || currentDir === this.config.stopDir) {
-      return null;
-    }
-    return nextDir;
-  }
-
-  private loadPackageProp(filepath: string, content: string): unknown {
-    const parsedContent = loaders.loadJson(filepath, content);
-    const packagePropValue = getPropertyByPath(
-      parsedContent,
-      this.config.packageProp,
-    );
-    return packagePropValue || null;
-  }
-
-  private getLoaderEntryForFile(filepath: string): LoaderEntry {
-    if (path.basename(filepath) === 'package.json') {
-      const loader = this.loadPackageProp.bind(this);
-      return { sync: loader, async: loader };
-    }
-
-    const loaderKey = path.extname(filepath) || 'noExt';
-    return this.config.loaders[loaderKey] || {};
-  }
-
   private getAsyncLoaderForFile(filepath: string): LoaderAsync {
     const entry = this.getLoaderEntryForFile(filepath);
     const loader = entry.async || entry.sync;
@@ -235,16 +77,6 @@ class Explorer {
       );
     }
     return loader;
-  }
-
-  private getSyncLoaderForFile(filepath: string): LoaderSync {
-    const entry = this.getLoaderEntryForFile(filepath);
-    if (!entry.sync) {
-      throw new Error(
-        `No sync loader specified for ${getExtensionDescription(filepath)}`,
-      );
-    }
-    return entry.sync;
   }
 
   private async loadFileContent(
@@ -262,34 +94,6 @@ class Explorer {
     return loaderResult;
   }
 
-  private loadFileContentSync(
-    filepath: string,
-    content: string | null,
-  ): LoadedFileContent {
-    if (content === null) {
-      return null;
-    }
-    if (content.trim() === '') {
-      return undefined;
-    }
-    const loader = this.getSyncLoaderForFile(filepath);
-    const loaderResult = loader(filepath, content);
-    return loaderResult;
-  }
-
-  private loadedContentToCosmiconfigResult(
-    filepath: string,
-    loadedContent: LoadedFileContent,
-  ): CosmiconfigResult {
-    if (loadedContent === null) {
-      return null;
-    }
-    if (loadedContent === undefined) {
-      return { filepath, config: undefined, isEmpty: true };
-    }
-    return { config: loadedContent, filepath };
-  }
-
   private async createCosmiconfigResult(
     filepath: string,
     content: string | null,
@@ -298,22 +102,6 @@ class Explorer {
     const result = this.loadedContentToCosmiconfigResult(filepath, fileContent);
 
     return result;
-  }
-
-  private createCosmiconfigResultSync(
-    filepath: string,
-    content: string | null,
-  ): CosmiconfigResult {
-    const fileContent = this.loadFileContentSync(filepath, content);
-    const result = this.loadedContentToCosmiconfigResult(filepath, fileContent);
-
-    return result;
-  }
-
-  private validateFilePath(filepath: string): void {
-    if (!filepath) {
-      throw new Error('load and loadSync must pass a non-empty string');
-    }
   }
 
   public async load(filepath: string): Promise<CosmiconfigResult> {
@@ -341,58 +129,19 @@ class Explorer {
 
     return runLoad();
   }
-
-  public loadSync(filepath: string): CosmiconfigResult {
-    this.validateFilePath(filepath);
-    const absoluteFilePath = path.resolve(process.cwd(), filepath);
-
-    const runLoadSync = (): CosmiconfigResult => {
-      const content = readFileSync(absoluteFilePath, { throwNotFound: true });
-      const cosmiconfigResult = this.createCosmiconfigResultSync(
-        absoluteFilePath,
-        content,
-      );
-
-      const transformResult = this.config.transform(cosmiconfigResult);
-
-      // @ts-ignore
-      return transformResult;
-    };
-
-    if (this.loadSyncCache) {
-      return cacheWrapperSync(
-        this.loadSyncCache,
-        absoluteFilePath,
-        runLoadSync,
-      );
-    }
-
-    return runLoadSync();
-  }
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 function createExplorer(options: ExplorerOptions) {
-  const explorer = new Explorer(options);
+  const explorer = new ExplorerAsync(options);
 
   return {
     search: explorer.search.bind(explorer),
-    searchSync: explorer.searchSync.bind(explorer),
     load: explorer.load.bind(explorer),
-    loadSync: explorer.loadSync.bind(explorer),
     clearLoadCache: explorer.clearLoadCache.bind(explorer),
     clearSearchCache: explorer.clearSearchCache.bind(explorer),
     clearCaches: explorer.clearCaches.bind(explorer),
   } as const;
-}
-
-function nextDirUp(dir: string): string {
-  return path.dirname(dir);
-}
-
-function getExtensionDescription(filepath: string): string {
-  const ext = path.extname(filepath);
-  return ext ? `extension "${ext}"` : 'files without extensions';
 }
 
 export { createExplorer };
